+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [2.0.3] - Renamed "limit" to "max"
+
+Current value and maximum now can be changed. Increment is now mutable.
+
+There is no "limit" now, there is a "max" representing maximum value increment's value can take. This should improve readability.
+
+Changes required if "limit" was utilized.
+
 ## [1.0.1] - Now supporting custom limit number
 
 ## [1.0.0] - Initial release